Fix dotnet/extensionsdotnet/extensionsdotnet/extensionsdotnet/extensions#8612 JSInterop throws truncated call stack
The fix for this is to use more ExceptionDispatchInfo! Basically
everwhere that we handle an exception is now an EDI. It's easy to pass
these around and they do the right thing as far as perserving the stack
in.

I de-factored this code a little bit to make all of this work, but it's
now pretty unambiguously correct upon inspection.

I started out wanting to get rid of the continue-with because we've
found that pretty error prone overall. However making this method
async-void started to ask more questions than it answered. What happens
if serialization of the exception fails?
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@dotnet/extensions@dotnet/extensions@0a27fd3ad6604ca1f1278db5e01224a29fcab00d
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@dotnet/extensions@718f70141e38034fd15bbf55c1b55557a372cd7b
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@3fd3304e2bf30f035155d6292ef10dedae83fd8d
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/637960a1002e56b4ae531f15b6df6a4864d414d7