Add AssemblyFixture to our test infra
This is a feature that we're using in Templates and Blazor E2E tests to manage selenium.

It's a general purpose kind of thing, so it makes sense to make it more general. This requires using the
`[assembly: TestFramework()]`.
Also fixed a bug here where this feature broke collection fixtures.
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@dotnet/extensions@dotnet/extensions@208d44a985fa4d8e13eb00515a44c5104c61b2b0
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@dotnet/extensions@669f58a992713a88afe7db7f9cb1c2cef9bb497e
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@e00119c9b6f2d93731fca845f3494e608d543d4d
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/9f5b361de1e00ec006c9878b98e50dfef0760c80