Add support for JSInvokable methods on generic types (dotnet/extensionsdotnet/extensionsdotnet/extensionsdotnet/extensions#2342)
* Add support for JSInvokable methods on generic types

Prior to this change, DotNetDispatcher cached the MethodInfo on the
generic type definition. Using this would have required MethodInfo.MakeGenericMethod before the method was invoked.
We could separately cache the result of this to avoid the reflection cost per invocation.

Alternatively we could cache static and non-static MethodInfo instances separately which is what this change attempts to do.
The big difference in the outcome is that this requires instance (non-static) JSInvokable methods to be only unique named within
the type hierarchy as opposed to across all static and instance JSInvokable methods in an assembly.

Fixes https://github.com/aspnet/Extensions/issues/1360
Fixes https://github.com/aspnet/AspNetCore/issues/9061
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@dotnet/extensions@dotnet/extensions@659b604fb2e595d48a931a7ed831f6c38f035382
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@dotnet/extensions@aaf9159b65e6fd911ffc63514bc3b179d4122d08
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/dotnet/extensions@4f70b0248ef2ba6547cbe93bdea8169c2b826898
\n\nCommit migrated from https://github.com/dotnet/extensions/commit/92c60ffe133400d7756401b65a3a0123a0b340e6